Materialize TreeViewItems on selection.
If we're selecting a particular tree view item, then materialize and expand the item's ancestors as `SelectionModel` requests children. Only do this if a particular item is being selected, not when an item is selected as part of a range select.

To do this, needed to add a `FinalIndex`  property to `SelectionModelChildrenRequestedEventArgs` in order to know if we're selecting a descendent of the item whose children are being requested.

This is a massive hack, but I can't think of a better way to do it with the current `TreeView` implementation.